chatbot/requirements.txt
Flask==2.0.1
twilio==6.61.0
python-dotenv==0.19.0
